+++ src/modules/client/components/searchResultCard/ClientSearchResultCard.stories.tsx
